requirements.txt
spleeter-gpu==2.3.0
librosa==0.8.0
pydub==0.25.1